requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.6.1
easyocr==1.7.2
email_validator==2.2.0
filelock==3.16.1
flash==1.0.3
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-WTF==1.2.1
fonttools==4.54.1
fsspec==2024.10.0
git-filter-repo==2.45.0
h11==0.14.0
idna==3.8
imageio==2.36.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
phonenumbers==8.13.45
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.5.0
shapely==2.0.6
simple-websocket==1.1.0
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics-thop==2.0.11
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
